email_gpt_bot-main/requirements.txt
openai
markdown-it-py
yagmail
